refactor: creating repository listener class
